Dolphin - RsrSocket
The socket implementation in Dolphin was not ideal for use with RSR. The
distinction between ServerSocket2 and Socket2 resulted in a confused
prior implementation.

The ServerSocket2 class in Dolphin does not allow binding to a specific
IP address. Instead, you always bind to the wildcard address.

I wouldn't be able to reasonably wrap the default socket classes and
maintain the behavior of RsrSocket. RsrSocket is implemented against the underlying library used to call Socket functions.

RsrSocket stores a file descriptor used for each library call. The
library calls are configured to get RsrSocketTestCase methods to pass.
They may not be ideal. Since IPv6 is not tested directly, it may not
work or may work partially.

Many of the methods are inspired by the methods in Socket2 and
ServerSocket2.