RsrAcceptConnection - Cancel active #waitForConnection
RsrAcceptConnection now supports canceling an active wait for connection. When the connection is cancelled, the #waitForConnection method will signal an RsrWaitForConnectionCancelled error.

Previously, there was not way to cancel an AcceptConnection. You could kill the process but the listening socket would remain open for some period of time. The listening socket will now be closed and the wait will terminate.