RetainObject - Removal
RetainObject does not make much sense as an independent command. Treating it separately results in the need to cache objects in in the Connection until a future SendMessage or DeliverResult message are received.

It makes more sense to embed all necessary services into SendMessage or DeliverResult. This assigns direct relation vs infered relation based upon having received a Service shortly beforehand.

The tests have been cleaned up or simplified in response to the removal of RetainObject and ObjectCache.