Removed Command>>#sendOver:
This method pushed code around in a less than ideal way. SendMessage was responsible for updating data structures in Connection. #sendOver: also previously ensured the Command was encoded to bytes.

Encoding is a detail of how Command are sent from one Connection to another. They do not belong in the high-level abstraction. Instead, Connection currently creates the encoding and then assigns it in the Command instance. Later, the encoding will happen just before writing to the socket.