Encode Commands directly to SocketStream
Command encoding now occurs directly to the SocketStream. This does mean an error will need to terminate the connection. If the connection is not terminated, unknown behavior could occur. This should be minimized by proper error handling when creating a Snapshot of the transitive closure for a message send or response.