Service - #preUpdate and #postUpdate
RSR automatically updates instance variables in Services. Previously, the Service was not notified that this will or has occurred. It is possible that the service would need to stash temporary objects or changes before the update is applied.

This change adds #preUpdate and #postUpdate messages to ensure Services know their update windows. This processing happens depth-first to ensure child Services are already stable during the call.