Server instance lifetimes
This commit updates the tests which verify a Server's lifetime. The tests are a bit more direct now in their assertions. The tests failed in GemStone before fixing. Pharo showed lifetime issues with normal tests though they were not consistent.

Previous to this commit, it appeared that a Server could be removed from the registry before the associated Client was garbage collected. The additional tests explicitely test the lifetime of Services in a way that should fail if lifetime handling is incorrect.