Async sends to #remoteSelf
Message sends to #remoteSelf are now asynchronous. The message send will return 'immediately' and the result will be a Promise. Eventually, the Promise will be resolved. The resolution value can be accessed by sending #wait or #value. Both methods will wait until the Promise is resolved before returning.

If the Promise is broken, #wait/#value will raise an BrokenPromise exception.