BinaryStreamChannel
Previously, this class was named SocketChannel. This wasn't that greate of a name as the Channel only cared about the Socket so it could access a SocketStream. I now inject the SocketStream as both an 'inStream' and an 'outStream'. Made the separation explicite and updated workers to call #inStream and #outStream rather than a generic #stream.