Updated Forwarder to subclass Object.
Forwarder has caused a lot of problems with debuggers. Since it attempts to forward messages remotely, it returns a Promise to the debugger which causes it to blow up. This makes debugging RSR or apps that use it tricky.

Subclassing object does mean that users will need to be more careful about the selectors they send to #remoteSelf. If they aren't careful, it can result they may get a message implemented in Object rather than having a message go to the Client/Server on the other side of the Connection.