RsrAsyncMournHandler
The AsyncMournHandler class is a singleton which starts a process to mourn Ephemerons. This is necessary because Ephemeron instance mourning is an interrupt to the current thread otherwise. If RSR has entered the Registry's critical section, the mourning of Services will deadlock.

The #testEphemeronMourningDoesNotDeadlock test ensures that a locked Registry does not prevent mourning from occurring eventually.